requirements.txt
Flask>=3.0.0
google-generativeai>=0.3.0
PyYAML>=6.0
GitPython>=3.1.40
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-flask>=1.3.0
